decouple state from component visibility
