+++ app/packages/core/src/components/Schema/SchemaSearch.tsx
@@ -56,20 +56,22 @@ export const SchemaSearch = (props: Props) => {
                 finalSearchTerm = finalSearchTerm.trim();
                 checkValue = checkValue.trim();
 
-                let props = finalSearchTerm.split(".");
-                const last = props[props.length - 1];
-                let object: {} = { include_nested_fields: includeNestedFields };
+                const termSplit = finalSearchTerm.split(".");
+                const last = termSplit[termSplit.length - 1];
+                let object: object = {
+                  include_nested_fields: includeNestedFields,
+                };
                 let ref = object;
-                if (checkValue && props.length > 0) {
-                  props.forEach((prop, index) => {
-                    if (index === props.length - 1) {
+                if (checkValue && termSplit.length > 0) {
+                  termSplit.forEach((prop, index) => {
+                    if (index === termSplit.length - 1) {
                       ref[prop] = checkValue;
                     } else {
                       ref[prop] = {};
                     }
                     ref = ref[prop];
                   });
-                  props[last] = checkValue;
+                  termSplit[last] = checkValue;
                 } else {
                   object = {
                     ...object,
