+++ app/packages/state/src/hooks/index.ts
@@ -49,5 +49,4 @@ export { default as useTooltip } from "./useTooltip";
 export { default as useToPatches } from "./useToPatches";
 export { default as useUpdateSamples } from "./useUpdateSamples";
 export { default as useSchemaSettings } from "./useSchemaSettings";
-export { settingsModal } from "./useSchemaSettings";
 export { default as withSuspense } from "./withSuspense";
