+++ app/packages/state/src/hooks/useSchemaSettings.ts
@@ -1,18 +1,9 @@
 import * as foq from "@fiftyone/relay";
 import * as fos from "@fiftyone/state";
-import { buildSchema } from "@fiftyone/state";
-import {
-  DYNAMIC_EMBEDDED_DOCUMENT_FIELD_V2,
-  EMBEDDED_DOCUMENT_FIELD,
-  LIST_FIELD,
-  Schema,
-} from "@fiftyone/utilities";
 import { isEmpty, keyBy } from "lodash";
 import { useCallback, useContext, useEffect, useMemo } from "react";
 import { useMutation, useRefetchableFragment } from "react-relay";
 import {
-  atom,
-  atomFamily,
   useRecoilCallback,
   useRecoilState,
   useRecoilValue,
@@ -21,298 +12,11 @@ import {
 } from "recoil";
 import { disabledField, skipField } from "./useSchemaSettings.utils";
 
-export const TAB_OPTIONS_MAP = {
-  SELECTION: "Selection",
-  FILTER_RULE: "Filter rule",
-};
-
-export const TAB_OPTIONS = Object.values(TAB_OPTIONS_MAP);
 const SELECT_ALL = "SELECT_ALL";
 
-export const schemaSearchTerm = atom<string>({
-  key: "schemaSearchTerm",
-  default: "",
-});
-
-export const showNestedFieldsState = atom<boolean>({
-  key: "showNestedFieldsState",
-  default: false,
-});
-
-export const schemaSelectedSettingsTab = atom<string>({
-  key: "schemaSelectedSettingsTab",
-  default: TAB_OPTIONS_MAP.SELECTION,
-});
-
-export const settingsModal = atom<{ open: boolean } | null>({
-  key: "settingsModal",
-  default: {
-    open: false,
-  },
-});
-
-export const allFieldsCheckedState = atom<boolean>({
-  key: "allFieldsCheckedState",
-  default: true,
-});
-
-export const expandedPathsState = atom<{} | null>({
-  key: "expandedPathsState",
-  default: null,
-});
-
-export const schemaSearchRestuls = atom<string[]>({
-  key: "schemaSearchRestuls",
-  default: [],
-  effects: [
-    ({ onSet, getPromise, setSelf }) => {
-      onSet(async (newPaths = []) => {
-        const viewSchema = await getPromise(viewSchemaState);
-        const fieldSchema = await getPromise(fieldSchemaState);
-        const combinedSchema = { ...fieldSchema, ...viewSchema };
-
-        const greenPaths = [...newPaths]
-          .filter(
-            (path) =>
-              path &&
-              combinedSchema?.[
-                path.startsWith("frames.") ? path.replace("frames.", "") : path
-              ]?.ftype &&
-              !skipField(path, combinedSchema)
-          )
-          .map((path) =>
-            path.startsWith("frames.") ? path.replace("frames.", "") : path
-          );
-
-        setSelf(greenPaths);
-      });
-    },
-  ],
-});
-
-// tracks action and value - currently used for (un)select all action
-export const lastActionToggleSelectionState = atom<Record<
-  string,
-  boolean
-> | null>({
-  key: "lastActionToggleSelectionState",
-  default: null,
-});
-
-export const selectedPathsState = atomFamily({
-  key: "selectedPathsState",
-  default: (param: {}) => param,
-  effects: [
-    ({ onSet, getPromise, setSelf }) => {
-      onSet(async (newPathsMap) => {
-        const dataset = await getPromise(fos.dataset);
-        const viewSchema = await getPromise(viewSchemaState);
-        const fieldSchema = await getPromise(fieldSchemaState);
-        const combinedSchema = { ...fieldSchema, ...viewSchema };
-        const isImage = dataset.mediaType === "image";
-        const isVideo = dataset.mediaType === "video";
-
-        const mapping = {};
-        Object.keys(combinedSchema).forEach((path) => {
-          if (isImage) {
-            mapping[path] = path;
-          }
-          if (isVideo && viewSchema) {
-            Object.keys(viewSchema).forEach((path) => {
-              mapping[path] = `frames.${path}`;
-            });
-          }
-        });
-
-        const newPaths = newPathsMap?.[dataset?.name] || [];
-        const greenPaths = [...newPaths]
-          .filter((path) => {
-            const skip = skipField(path, combinedSchema);
-            return !!path && !skip;
-          })
-          .map((path) => mapping?.[path] || path);
-
-        setSelf({
-          [dataset?.name]: new Set(greenPaths),
-        });
-      });
-    },
-  ],
-});
-
-export const excludedPathsState = atomFamily({
-  key: "excludedPathsState",
-  default: (param: {}) => param,
-  effects: [
-    ({ onSet, getPromise, setSelf }) => {
-      onSet(async (newPathsMap) => {
-        const viewSchema = await getPromise(viewSchemaState);
-        const fieldSchema = await getPromise(fieldSchemaState);
-        const dataset = await getPromise(fos.dataset);
-        const showNestedField = await getPromise(showNestedFieldsState);
-        const searchResults = await getPromise(schemaSearchRestuls);
-        const isFrameVIew = await getPromise(fos.isFramesView);
-        const isVideo = dataset.mediaType === "video";
-        const isImage = dataset.mediaType === "image";
-        const isInSearchMode = !!searchResults?.length;
-
-        if (!dataset) {
-          return;
-        }
-
-        const combinedSchema = { ...fieldSchema, ...viewSchema };
-        const mapping = {};
-        Object.keys(combinedSchema).forEach((path) => {
-          if (isImage) {
-            mapping[path] = path;
-          }
-          if (isVideo && viewSchema) {
-            Object.keys(viewSchema).forEach((path) => {
-              mapping[path] = `frames.${path}`;
-            });
-          }
-        });
-
-        const newPaths = newPathsMap?.[dataset?.name] || [];
-        const greenPaths = [...newPaths]
-          .filter((path) => {
-            const rawPath = path.replace("frames.", "");
-            return (
-              !!rawPath &&
-              !skipField(rawPath, combinedSchema) &&
-              !disabledField(
-                path,
-                combinedSchema,
-                dataset?.groupField,
-                isFrameVIew
-              )
-            );
-          })
-          .map((path) => mapping?.[path] || path);
-
-        // if top level only, count should be top-level too
-        // if nested fields are shown, exclude more granular
-        let finalGreenPaths = greenPaths;
-        if (!showNestedField && !isInSearchMode) {
-          finalGreenPaths = greenPaths.filter((path) =>
-            isVideo
-              ? (path.split(".").length === 2 && path.startsWith("frames.")) ||
-                !path.includes(".")
-              : !path.includes(".")
-          );
-        }
-
-        const shouldFilterTopLevelFields = showNestedField || isInSearchMode;
-        finalGreenPaths = shouldFilterTopLevelFields
-          ? finalGreenPaths.filter((path) => {
-              const isEmbeddedOrListType = [
-                EMBEDDED_DOCUMENT_FIELD,
-                LIST_FIELD,
-              ].includes(combinedSchema[path]?.ftype);
-
-              // embedded document could break an exclude_field() call causing mongo query issue.
-              const hasDynamicEmbeddedDocument = [
-                DYNAMIC_EMBEDDED_DOCUMENT_FIELD_V2,
-              ].includes(combinedSchema[path]?.embeddedDocType);
-
-              const isTopLevelPath = isVideo
-                ? !(
-                    path.split(".").length === 2 && path.startsWith("frames.")
-                  ) || !path.includes(".")
-                : !path.includes(".");
-
-              return !(
-                isEmbeddedOrListType &&
-                hasDynamicEmbeddedDocument &&
-                isTopLevelPath
-              );
-            })
-          : finalGreenPaths;
-
-        // filter out subpaths if a parent (or higher) path is also in the list
-        const finalGreenPathsSet = new Set(finalGreenPaths);
-        if (showNestedField) {
-          finalGreenPaths = finalGreenPaths.filter((path) => {
-            let tmp = path;
-            while (tmp.indexOf(".") > 0) {
-              const parentPath = tmp.substring(0, tmp.lastIndexOf("."));
-              if (finalGreenPathsSet.has(parentPath) || path === parentPath) {
-                return false;
-              }
-              tmp = parentPath;
-            }
-            return true;
-          });
-        }
-
-        setSelf({
-          [dataset.name]: new Set(finalGreenPaths),
-        });
-      });
-    },
-  ],
-});
-export const schemaState = atom<Schema>({
-  key: "schemaState",
-  default: null,
-});
-
-export const viewSchemaState = atom({
-  key: "viewSchemaState",
-  default: null,
-});
-export const fieldSchemaState = atom({
-  key: "fieldSchemaState",
-  default: null,
-});
-export const showMetadataState = atom({
-  key: "showMetadataState",
-  default: false,
-});
-export const includeNestedFieldsState = atom({
-  key: "includeNestedFieldsState",
-  default: true,
-});
-export const affectedPathCountState = atom({
-  key: "affectedPathCountState",
-  default: 0,
-});
-export const searchMetaFilterState = atom({
-  key: "searchMetaFilterState",
-  default: {},
-});
-export const lastAppliedPathsState = atom({
-  key: "lastAppliedExcludedPathsState",
-  default: {
-    excluded: [],
-    selected: [],
-  },
-});
-
-export const selectedFieldsStageState = atom<any>({
-  key: "selectedFieldsStageState",
-  default: undefined,
-  effects: [
-    ({ onSet }) => {
-      onSet((value) => {
-        const context = fos.getContext();
-        if (context.loaded) {
-          context.history.replace(
-            `${context.history.location.pathname}${context.history.location.search}`,
-            {
-              ...context.history.location.state,
-              selectedFieldsStage: value || null,
-            }
-          );
-        }
-      });
-    },
-  ],
-});
-
 export default function useSchemaSettings() {
-  const [settingModal, setSettingsModal] = useRecoilState(settingsModal);
-  const [showMetadata, setShowMetadata] = useRecoilState(showMetadataState);
+  const [settingModal, setSettingsModal] = useRecoilState(fos.settingsModal);
+  const [showMetadata, setShowMetadata] = useRecoilState(fos.showMetadataState);
   const router = useContext(fos.RouterContext);
   const [setView] = useMutation<foq.setViewMutation>(foq.setView);
   const dataset = useRecoilValue(fos.dataset);
@@ -321,7 +25,7 @@ export default function useSchemaSettings() {
   const resetTextFilter = useResetRecoilState(fos.textFilter(false));
   const datasetName = useRecoilValue(fos.datasetName);
 
-  const resetSelectedPaths = useResetRecoilState(selectedPathsState({}));
+  const resetSelectedPaths = useResetRecoilState(fos.selectedPathsState({}));
 
   const setSelectedFieldsStage = useRecoilCallback(
     ({ snapshot, set }) =>
@@ -330,7 +34,7 @@ export default function useSchemaSettings() {
           return;
         }
 
-        set(selectedFieldsStageState, value);
+        set(fos.selectedFieldsStageState, value);
 
         // router is loaded only in OSS
         if (router.loaded) return;
@@ -356,50 +60,49 @@ export default function useSchemaSettings() {
     [setView, router, dataset]
   );
 
-  const setViewSchema = useSetRecoilState(viewSchemaState);
-  const setFieldSchema = useSetRecoilState(fieldSchemaState);
-  const [searchTerm, setSearchTerm] = useRecoilState<string>(schemaSearchTerm);
-  const [searchResults, setSearchResults] = useRecoilState(schemaSearchRestuls);
+  const setViewSchema = useSetRecoilState(fos.viewSchemaState);
+  const setFieldSchema = useSetRecoilState(fos.fieldSchemaState);
+  const [searchTerm, setSearchTerm] = useRecoilState<string>(
+    fos.schemaSearchTerm
+  );
+  const [searchResults, setSearchResults] = useRecoilState(
+    fos.schemaSearchResults
+  );
   const isVideo = dataset.mediaType === "video";
 
-  const setSchema = useSetRecoilState(schemaState);
-  useEffect(() => {
-    if (datasetName) {
-      setSchema(dataset ? buildSchema(dataset, true) : null);
-    }
-  }, [datasetName]);
-
   const [allFieldsChecked, setAllFieldsChecked] = useRecoilState(
-    allFieldsCheckedState
+    fos.allFieldsCheckedState
   );
 
   const [includeNestedFields, setIncludeNestedFieldsRaw] = useRecoilState(
-    includeNestedFieldsState
+    fos.includeNestedFieldsState
   );
 
   const [affectedPathCount, setAffectedPathCount] = useRecoilState(
-    affectedPathCountState
+    fos.affectedPathCountState
   );
 
   const [lastAppliedPaths, setLastAppliedPaths] = useRecoilState(
-    lastAppliedPathsState
+    fos.lastAppliedPathsState
   );
 
   const [showNestedFields, setShowNestedFieldsRaw] = useRecoilState<boolean>(
-    showNestedFieldsState
+    fos.showNestedFieldsState
   );
 
   const [searchMetaFilter, setSearchMetaFilter] = useRecoilState(
-    searchMetaFilterState
+    fos.searchMetaFilterState
   );
 
   const isPatchesView = useRecoilValue(fos.isPatchesView);
   const isFrameView = useRecoilValue(fos.isFramesView);
 
-  const [expandedPaths, setExpandedPaths] = useRecoilState(expandedPathsState);
+  const [expandedPaths, setExpandedPaths] = useRecoilState(
+    fos.expandedPathsState
+  );
 
   const [lastActionToggleSelection, setLastActionToggleSelection] =
-    useRecoilState(lastActionToggleSelectionState);
+    useRecoilState(fos.lastActionToggleSelectionState);
 
   const vStages = useRecoilValue(fos.view);
   const [data, refetch] = useRefetchableFragment<
@@ -441,11 +144,11 @@ export default function useSchemaSettings() {
   }, [viewSchema, fieldSchema]);
 
   const [selectedTab, setSelectedTab] = useRecoilState(
-    schemaSelectedSettingsTab
+    fos.schemaSelectedSettingsTab
   );
-  const filterRuleTab = selectedTab === TAB_OPTIONS_MAP.FILTER_RULE;
+  const filterRuleTab = selectedTab === fos.TAB_OPTIONS_MAP.FILTER_RULE;
 
-  const selectedPathState = selectedPathsState({});
+  const selectedPathState = fos.selectedPathsState({});
   const [selectedPaths, setSelectedPaths] = useRecoilState<{}>(
     selectedPathState
   );
@@ -460,7 +163,7 @@ export default function useSchemaSettings() {
         )
       : [];
 
-  const excludePathsState = excludedPathsState({});
+  const excludePathsState = fos.excludedPathsState({});
   const [excludedPaths, setExcludedPaths] = useRecoilState<{}>(
     excludePathsState
   );
