+++ app/packages/state/src/hooks/useSchemaSettings.utils.test.ts
@@ -7,8 +7,6 @@ import {
   DETECTIONS_FIELD,
   DETECTION_DISABLED_SUB_PATHS,
   DETECTION_FIELD,
-  DYNAMIC_EMBEDDED_DOCUMENT_FIELD,
-  DYNAMIC_EMBEDDED_DOCUMENT_FIELD_V2,
   EMBEDDED_DOCUMENT_FIELD,
   FLOAT_FIELD,
   FRAME_NUMBER_FIELD,
@@ -25,8 +23,6 @@ import {
   REGRESSION_FIELD,
   SEGMENTATION_FIELD,
   STRING_FIELD,
-  TEMPORAL_DETECTION,
-  TEMPORAL_DETECTIONS,
   TEMPORAL_DETECTIONS_FIELD,
   TEMPORAL_DETECTION_FIELD,
   CLASSIFICATION_DISABLED_SUB_PATHS,
