+++ app/packages/state/src/recoil/index.ts
@@ -16,3 +16,4 @@ export * from "./sidebar";
 export * from "./types";
 export * from "./utils";
 export * from "./view";
+export * from "./schemaSettings.atoms";
